Update ServerDescription to add BlobUploadLimit. See https://github.com/bluesky-social/atproto/pull/5277
