+++ src/idunno.AtProto/Server/Defs/Links.cs
@@ -1,0 +1,24 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace idunno.AtProto.Server;
+
+/// <summary>
+/// Links a server may provide as part its description.
+/// </summary>
+public sealed record Links
+{
+    /// <summary>
+    /// Gets a URI to the server's privacy policy.
+    /// </summary>
+    [JsonInclude]
+    public Uri? PrivacyPolicy { get; init; }
+
+    /// <summary>
+    /// Gets a URI to the server's terms of service.
+    /// </summary>
+    [JsonInclude]
+    public Uri? TermsOfService { get; init; }
+}
